Renderiza loading ao carregar a pagina
