+++ exercises/bloco_13/random-user-app/src/components/Loading.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Loading extends Component {
+  render() {
+    return (
+      <div>
+        <p>Loading...</p>
+      </div>
+    );
+  }
+}
+
+export default Loading;
