+++ exercises/bloco_13/random-user-app/src/components/PersonalData.jsx
@@ -1,39 +1,18 @@
 import React, { Component } from 'react';
 
 class PersonalData extends Component {
-  constructor() {
-    super();
-
-    this.state = {
-      users: [],
-    }
-
-    this.fetchUsers = this.fetchUsers.bind(this);
-  }
-
-  async fetchUsers() {
-    const requestHeaders = { headers: { Accept: 'aplication/json' } }
-    const requestReturn = await fetch('https://api.randomuser.me/', requestHeaders);
-    const requestObject = await requestReturn.json();
-    const results = await requestObject.results;
-
-    this.setState({
-      users: results,
-    })
-  }
-
-  componentDidMount() {
-    this.fetchUsers();
-  }
-
   render() {
-    const { users } = this.state;
+    const { users } = this.props;
 
     return (
       <div>
         { users.map((user) => (
           <div key={ user.login.uuid }>
+            <img src={user.picture.large} alt="random user profile" />
             <p>{ user.name.first }</p>
+            <p>{ user.email }</p>
+            <p>{ user.dob.age }</p>
+            <p>{ user.location.country }</p>
           </div>
         )) }
       </div>
